Refactor PyWhoIs
The PyWhoIs class was replaced with a more OOP friendly approach like the one in query.py. The DomainLookup subclass contains the logic related to performing domain lookups for both WHOIS and RDAP queries. The NumberLookup class was added to implement the logic necessary for ipv4 adn ipv4 queries (and RDAP ASN).